flask/requirements.txt
flask==2.1.0
opencv-python-headless==4.5.5.64
numpy>=1.23.5,<2.0.0
Pillow==9.4.0
google-generativeai==0.3.1
werkzeug==2.1.1
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
tensorflow==2.13.0
keras==2.13.1
grpcio-status==1.51.1
